requirements.txt
asgiref==3.5.0
bcrypt==3.2.0
cffi==1.15.0
colorama==0.4.4
Django==4.0.4
django-cors-headers==3.11.0
django-filter==21.1
django-jazzmin==2.5.0
django-rest-auth==0.9.5
djangorestframework==3.13.1
djangorestframework-simplejwt==5.1.0
Markdown==3.3.6
Pillow==9.1.0
psycopg2==2.9.3
pycparser==2.21
PyJWT==2.3.0
pytz==2022.1
six==1.16.0
sqlparse==0.4.2
tzdata==2022.1
django-dotenv
